Refactor monitor parsing into a specific module
