+++ pyFAI/test/test_all.py
@@ -33,7 +33,7 @@ __authors__ = ["Jérôme Kieffer"]
 __contact__ = "jerome.kieffer@esrf.eu"
 __license__ = "MIT"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "04/01/2019"
+__date__ = "19/02/2019"
 
 import sys
 import unittest
@@ -76,6 +76,7 @@ from . import test_worker
 from . import test_utils_shell
 from . import test_utils_stringutil
 from . import test_utils_mathutil
+from . import test_utils_header
 from . import test_preproc
 from . import test_bayes
 from . import test_scripts
@@ -130,6 +131,7 @@ def suite():
     testsuite.addTest(test_utils_shell.suite())
     testsuite.addTest(test_utils_stringutil.suite())
     testsuite.addTest(test_utils_mathutil.suite())
+    testsuite.addTest(test_utils_header.suite())
     testsuite.addTest(test_preproc.suite())
     testsuite.addTest(test_bayes.suite())
     testsuite.addTest(test_scripts.suite())
